Move SourceReader to warehouse.py
